+++ app/src/main/java/com/dunbar/termtracker/CourseDetailActivity.java
@@ -8,9 +8,11 @@ import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.List;
 
 public class CourseDetailActivity extends AppCompatActivity {
     dbHelper db;
@@ -18,10 +20,12 @@ public class CourseDetailActivity extends AppCompatActivity {
     EditText txtStartDate;
     EditText txtEndDate;
     Spinner courseStatusDropdown;
+    Spinner alertOptionDropdown;
     Course course;
     Intent intent;
     SimpleDateFormat formatter;
     ArrayAdapter spinnerAdapter;
+    ArrayAdapter alertSpinnerAdapter;
     int termId;
 
     @Override
@@ -39,6 +43,7 @@ public class CourseDetailActivity extends AppCompatActivity {
         txtStartDate = (EditText)findViewById(R.id.txtCourseStartDate);
         txtEndDate = (EditText)findViewById(R.id.txtCourseEndDate);
         courseStatusDropdown = (Spinner)findViewById(R.id.courseStatusDropdown);
+        alertOptionDropdown = (Spinner)findViewById(R.id.courseAlertOptionDropdown);
         String[] statuses = getResources().getStringArray(R.array.status_array);
         spinnerAdapter = new ArrayAdapter<String>(
                 this,
@@ -46,6 +51,13 @@ public class CourseDetailActivity extends AppCompatActivity {
                 statuses
                 );
         courseStatusDropdown.setAdapter(spinnerAdapter);
+        String[] alertOptions = getResources().getStringArray(R.array.alert_type_array);
+        alertSpinnerAdapter = new ArrayAdapter<String>(
+                this,
+                android.R.layout.simple_spinner_dropdown_item,
+                alertOptions
+        );
+        alertOptionDropdown.setAdapter(alertSpinnerAdapter);
 
         if(intent.getExtras().getBoolean("isNew")){
             createNewCourse();
@@ -67,6 +79,13 @@ public class CourseDetailActivity extends AppCompatActivity {
         txtEndDate.setText(formatter.format(course.getEndDate()));
         int spinnerPosition = spinnerAdapter.getPosition(course.getStatus());
         courseStatusDropdown.setSelection(spinnerPosition);
+        int alertSpinnerPosition = 0;
+        if(course.getAlert()){
+            alertSpinnerPosition = alertSpinnerAdapter.getPosition("ON");
+        }else{
+            alertSpinnerPosition = alertSpinnerAdapter.getPosition("OFF");
+        }
+        alertOptionDropdown.setSelection(alertSpinnerPosition);
     }
 
     public void saveCourse(View view){
@@ -77,7 +96,11 @@ public class CourseDetailActivity extends AppCompatActivity {
             course.setStartDate(formatter.parse(txtStartDate.getText().toString()));
             course.setEndDate(formatter.parse(txtEndDate.getText().toString()));
             course.setStatus(courseStatusDropdown.getSelectedItem().toString());
-            course.setAlert(false); //TODO - CHANGE THIS!!!!!!!!
+            if(alertOptionDropdown.getSelectedItem().toString().equals("ON")){
+                course.setAlert(true);
+            }else{
+                course.setAlert(false);
+            }
         } catch (ParseException e) {
             e.printStackTrace();
         }
@@ -94,27 +117,62 @@ public class CourseDetailActivity extends AppCompatActivity {
     }
 
     public void deleteCourse(View view){
-        db.deleteCourse(course.getId());
-        goBack();
+        //show error if course has not been saved
+        if(course.getId() == 0){
+            Toast.makeText(getApplicationContext(),"You cannot delete a course that has not been saved.",Toast.LENGTH_LONG).show();
+            return;
+        }
+
+        //show error if the course has any assessments, mentors or notes associated with it
+        List<Assessment> assessments = db.getAssessments(course.getId());
+        List<Mentor> mentors = db.getMentors(course.getId());
+        List<CourseNotes> courseNotes = db.getCourseNotesByCourseId(course.getId());
+        if(assessments.size() > 0 || mentors.size() > 0 || courseNotes.size() > 0){
+            Toast.makeText(getApplicationContext(),"You cannot delete a course that has assessments, mentors or notes associated with it.",Toast.LENGTH_LONG).show();
+            return;
+        }else{
+            db.deleteCourse(course.getId());
+            goBack();
+        }
     }
 
     public void viewMentors(View view){
+        //show error if course has not been saved
+        if(course.getId() == 0){
+            Toast.makeText(getApplicationContext(),"You cannot create mentors until the course has been saved.",Toast.LENGTH_LONG).show();
+            return;
+        }
         Intent intent = new Intent(this, MentorListActivity.class);
         intent.putExtra("courseId",course.getId());
         startActivity(intent);
     }
     public void viewAssessments(View view){
+        //show error if course has not been saved
+        if(course.getId() == 0){
+            Toast.makeText(getApplicationContext(),"You cannot create assessments until the course has been saved.",Toast.LENGTH_LONG).show();
+            return;
+        }
         Intent intent = new Intent(this, AssessmentListActivity.class);
         intent.putExtra("courseId",course.getId());
         startActivity(intent);
     }
     public void viewCourseNotes(View view){
+        //show error if course has not been saved
+        if(course.getId() == 0){
+            Toast.makeText(getApplicationContext(),"You cannot create course notes until the course has been saved.",Toast.LENGTH_LONG).show();
+            return;
+        }
         Intent intent = new Intent(this, CourseNotesListActivity.class);
         intent.putExtra("courseId",course.getId());
         startActivity(intent);
     }
 
     public void shareCourse(View view){
+        //show error if course has not been saved
+        if(course.getId() == 0){
+            Toast.makeText(getApplicationContext(),"You cannot share until the course has been saved.",Toast.LENGTH_LONG).show();
+            return;
+        }
         SimpleDateFormat formatter = new SimpleDateFormat("MM/dd/yyyy");
         Intent intent = new Intent(android.content.Intent.ACTION_SEND);
         intent.setType("text/plain");
